# loms/api/openapiv2/loms.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Loms",
    "version": "1.0.0"
  },
  "tags": [
    {
      "name": "Loms"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/order/cancel": {
      "post": {
        "operationId": "Loms_OrderCancel",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/lomsOrderCancelRequest"
            }
          }
        ],
        "tags": [
          "Loms"
        ]
      }
    },
    "/order/create": {
      "post": {
        "operationId": "Loms_OrderCreate",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/lomsOrderCreateResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/lomsOrderCreateRequest"
            }
          }
        ],
        "tags": [
          "Loms"
        ]
      }
    },
    "/order/info": {
      "post": {
        "operationId": "Loms_OrderInfo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/lomsOrderInfoResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/lomsOrderInfoRequest"
            }
          }
        ],
        "tags": [
          "Loms"
        ]
      }
    },
    "/order/info/all": {
      "get": {
        "operationId": "Loms_OrderInfoAll",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/lomsOrderInfoAllResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "Loms"
        ]
      }
    },
    "/order/pay": {
      "post": {
        "operationId": "Loms_OrderPay",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/lomsOrderPayRequest"
            }
          }
        ],
        "tags": [
          "Loms"
        ]
      }
    },
    "/stock/info": {
      "post": {
        "operationId": "Loms_StockInfo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/lomsStockInfoResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/lomsStockInfoRequest"
            }
          }
        ],
        "tags": [
          "Loms"
        ]
      }
    }
  },
  "definitions": {
    "OrderInfoResponseStatusEnum": {
      "type": "string",
      "enum": [
        "new",
        "awaiting_payment",
        "failed",
        "paid",
        "cancelled"
      ],
      "default": "new"
    },
    "lomsOrderCancelRequest": {
      "type": "object",
      "properties": {
        "orderId": {
          "type": "string",
          "format": "int64"
        },
        "userId": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "lomsOrderCreateRequest": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string",
          "format": "int64"
        },
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/lomsOrderItemCreateRequest"
          }
        }
      }
    },
    "lomsOrderCreateResponse": {
      "type": "object",
      "properties": {
        "orderId": {
          "type": "string",
          "format": "int64"
        },
        "userId": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "lomsOrderInfoAllResponse": {
      "type": "object",
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/lomsOrderInfoResponse"
          }
        }
      }
    },
    "lomsOrderInfoRequest": {
      "type": "object",
      "properties": {
        "orderId": {
          "type": "string",
          "format": "int64"
        },
        "userId": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "lomsOrderInfoResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "status": {
          "$ref": "#/definitions/OrderInfoResponseStatusEnum"
        },
        "user": {
          "type": "string",
          "format": "int64"
        },
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/lomsOrderItemInfoResponse"
          }
        }
      }
    },
    "lomsOrderItemCreateRequest": {
      "type": "object",
      "properties": {
        "sku": {
          "type": "integer",
          "format": "int64"
        },
        "count": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "lomsOrderItemInfoResponse": {
      "type": "object",
      "properties": {
        "sku": {
          "type": "integer",
          "format": "int64"
        },
        "count": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "lomsOrderPayRequest": {
      "type": "object",
      "properties": {
        "orderId": {
          "type": "string",
          "format": "int64"
        },
        "userId": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "lomsStockInfoRequest": {
      "type": "object",
      "example": {
        "sku": 773297411
      },
      "properties": {
        "sku": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "lomsStockInfoResponse": {
      "type": "object",
      "properties": {
        "count": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}